@@ -91,15 +91,16 @@ export default {
       API: '',
       senadores: [],
       filtro: {
-        periodo: 56,
+        periodo: 57,
         estado: [],
         partido: [],
       },
       periodos: [
-        { id: 56, text: '56º (2019-2023)' },
-        { id: 55, text: '55º (2015-2019)' },
-        { id: 54, text: '54º (2011-2015)' },
-        { id: 53, text: '53º (2007-2011)' },
+        { id: 57, text: '57ª (fev/2023 à jan/2027)' },
+        { id: 56, text: '56ª (fev/2019 à jan/2023)' },
+        { id: 55, text: '55ª (fev/2015 à jan/2019)' },
+        { id: 54, text: '54ª (fev/2011 à jan/2015)' },
+        { id: 53, text: '53ª (fev/2007 à jan/2011)' },
       ],
       estados: [],
       partidos: [],
