@@ -72,7 +72,7 @@
                     </div>
                     <div class="card-body">
                         <div class="table-responsive">
-                            <table class="table table-striped table-hover table-sm" style="margin: 0;">
+                            <table class="table table-striped table-hover table-sm" style="margin: 0;" aria-label="">
                                 <thead>
                                     <tr>
                                         <th style="width:80%">Fornecedor</th>
@@ -102,7 +102,7 @@
                     </div>
                     <div class="card-body">
                         <div class="table-responsive">
-                            <table class="table table-striped table-hover table-sm" style="margin: 0;">
+                            <table class="table table-striped table-hover table-sm" style="margin: 0;" aria-label="">
                                 <thead>
                                     <tr>
                                         <th style="width:80%">Fornecedor</th>
