@@ -7,7 +7,8 @@
         <div class="form-group col-md-2">
           <label>Ano</label>
           <select class="form-control input-sm" v-model="filtro.ano">
-            <option value="2022" selected>2022</option>
+            <option value="2023" selected>2023</option>
+            <option value="2022">2022</option>
             <option value="2021">2021</option>
             <option value="2020">2020</option>
             <option value="2019">2019</option>
@@ -125,15 +126,6 @@
       </div>
     </form>
 
-    <div class="row">
-      <div class="col-md-12">
-        <div class="alert alert-warning" v-if="valorTotal">
-          <strong>Custo Total no Período: R$ {{valorTotal}}</strong>
-          <small class="help-block mb-0">&nbsp;Custo Total considerando os filtros aplicados acima</small>
-        </div>
-      </div>
-    </div>
-
     <div class="form-group" v-if="fields">
       <vdtnet-table ref="table" :fields="fields" :opts="options" @edit="AbrirModalDetalhar"></vdtnet-table>
     </div>
@@ -187,8 +179,8 @@ export default {
       selectedRow: {},
       valorTotal: null,
       filtro: {
-        agrupar: '1',
-        ano: '2022',
+        agrupar: '7',
+        ano: '2023',
         mes: '',
         lotacao: [],
         cargo: [],
@@ -259,6 +251,7 @@ export default {
         serverSide: true,
         fixedHeader: true,
         saveState: true,
+        order: [],
       },
       fields: {},
     };
@@ -266,8 +259,8 @@ export default {
   mounted() {
     const vm = this;
 
-    vm.filtro.agrupar = vm.qs.ag || '1';
-    vm.filtro.ano = vm.qs.an || '2022';
+    vm.filtro.agrupar = vm.qs.ag || '7';
+    vm.filtro.ano = vm.qs.an || '2023';
     vm.filtro.mes = vm.qs.ms || '';
     vm.filtro.lotacao = (vm.qs.lt ? vm.qs.lt.split(',') : []);
     vm.filtro.cargo = (vm.qs.cr ? vm.qs.cr.split(',') : []);
@@ -335,8 +328,14 @@ export default {
     Pesquisar(pageLoad) {
       const vm = this;
       vm.senador = {};
-      vm.fields = null;
       vm.pageLoad = pageLoad;
+
+      if(vm.ultimoAgrupamento == vm.filtro.agrupar) {
+        vm.$refs.table.reload();
+        return;
+      }
+
+      vm.fields = null;
 
       vm.$nextTick(() => {
         switch (vm.filtro.agrupar) {
@@ -366,6 +365,7 @@ export default {
                 label: 'Custo Total',
                 sortable: true,
                 className: 'text-right',
+                defaultOrder: 'desc',
               },
             };
             break;
@@ -396,6 +396,7 @@ export default {
                 label: 'Custo Total',
                 sortable: true,
                 className: 'text-right',
+                defaultOrder: 'desc',
               },
             };
             break;
@@ -425,6 +426,7 @@ export default {
                 label: 'Custo Total',
                 sortable: true,
                 className: 'text-right',
+                defaultOrder: 'desc',
               },
             };
             break;
@@ -454,6 +456,7 @@ export default {
                 label: 'Custo Total',
                 sortable: true,
                 className: 'text-right',
+                defaultOrder: 'desc',
               },
             };
             break;
@@ -483,6 +486,7 @@ export default {
                 label: 'Custo Total',
                 sortable: true,
                 className: 'text-right',
+                defaultOrder: 'desc',
               },
             };
             break;
@@ -512,6 +516,7 @@ export default {
                 label: 'Custo Total',
                 sortable: true,
                 className: 'text-right',
+                defaultOrder: 'desc',
               },
             };
             break;
@@ -570,6 +575,11 @@ export default {
           default: break;
         }
 
+        vm.ultimoAgrupamento = vm.filtro.agrupar;
+        if(!pageLoad && vm.options.order.length > 0){
+          vm.options.order = [];
+        }
+
         vm.$nextTick(() => {
           vm.$refs.table.reload(null, true);
         });
@@ -578,7 +588,7 @@ export default {
     LimparFiltros() {
       this.filtro = {
         agrupar: '1',
-        ano: '2022',
+        ano: '2023',
         mes: '',
         lotacao: [],
         cargo: [],
